Model linearization with respect to epsilon.

The function performs a linearization of the model with respect to the residual variability. Derivative of model w.r.t. eps evaluated at eps=0

LinMatrixH(model_switch, xt_ind, x, a, bpop, b_ind, bocc_ind, poped.db)

Arguments

model_switch
A matrix that is the same size as xt, specifying which model each sample belongs to.

xt_ind
A vector of the individual/group sample times

x
A matrix for the discrete design variables. Each row is a group.

a
A matrix of covariates. Each row is a group.

bpop
The fixed effects parameter values. Supplied as a vector.

b_ind
vector of individual realization of the BSV terms b

bocc_ind
Vector of individual realizations of the BOV terms bocc

poped.db
A PopED database.

Value

A matrix of size (samples per individual x number of epsilons)

See also