Model linearization with respect to occasion variablity parameters.

The function performs a linearization of the model with respect to the occation variability parameter.. Derivative of model w.r.t. eta_occ, evaluated bocc_ind.

LinMatrixL_occ(model_switch, xt_ind, x, a, bpop, b_ind, bocc_ind, iCurrentOcc,
  poped.db)

Arguments

model_switch

A matrix that is the same size as xt, specifying which model each sample belongs to.

xt_ind

A vector of the individual/group sample times

x

A matrix for the discrete design variables. Each row is a group.

a

A matrix of covariates. Each row is a group.

bpop

The fixed effects parameter values. Supplied as a vector.

b_ind

vector of individual realization of the BSV terms b

bocc_ind

Vector of individual realizations of the BOV terms bocc

iCurrentOcc

The current occasion.

poped.db

A PopED database.

Value

A matrix of size (samples per individual x number of iovs)

See also

Other FIM: LinMatrixH, LinMatrixLH, calc_ofv_and_fim, ed_laplace_ofv, ed_mftot, efficiency, evaluate.e.ofv.fim, evaluate.fim, gradf_eps, mf3, mf5, mf6, mf7, mf8, mftot0, mftot1, mftot2, mftot3, mftot4, mftot5, mftot6, mftot7, mftot, mf, ofv_criterion, ofv_fim

Examples

library(PopED)

############# START #################
## Create PopED database
## (warfarin model for optimization)
#####################################

## Warfarin example from software comparison in:
## Nyberg et al., "Methods and software tools for design evaluation 
##   for population pharmacokinetics-pharmacodynamics studies", 
##   Br. J. Clin. Pharm., 2014. 

## Optimization using an additive + proportional reidual error  
## to avoid sample times at very low concentrations (time 0 or very late samples).

## find the parameters that are needed to define from the structural model
ff.PK.1.comp.oral.sd.CL

#> function(model_switch,xt,parameters,poped.db){
#>   ##-- Model: One comp first order absorption
#>   with(as.list(parameters),{
#>     y=xt
#>     y=(DOSE*Favail*KA/(V*(KA-CL/V)))*(exp(-CL/V*xt)-exp(-KA*xt))
#>     return(list( y= y,poped.db=poped.db))
#>   })
#> }
#> <environment: namespace:PopED>

## -- parameter definition function 
## -- names match parameters in function ff
sfg <- function(x,a,bpop,b,bocc){
  parameters=c(CL=bpop[1]*exp(b[1]),
               V=bpop[2]*exp(b[2]),
               KA=bpop[3]*exp(b[3]),
               Favail=bpop[4],
               DOSE=a[1])
  return(parameters)
}

## -- Define initial design  and design space
poped.db <- create.poped.database(ff_file="ff.PK.1.comp.oral.sd.CL",
                                  fg_file="sfg",
                                  fError_file="feps.add.prop",
                                  bpop=c(CL=0.15, V=8, KA=1.0, Favail=1),
                                  notfixed_bpop=c(1,1,1,0),
                                  d=c(CL=0.07, V=0.02, KA=0.6),
                                  sigma=c(0.01,0.25),
                                  groupsize=32,
                                  xt=c( 0.5,1,2,6,24,36,72,120),
                                  minxt=0.01,
                                  maxxt=120,
                                  a=70,
                                  mina=0.01,
                                  maxa=100)

#> Warning: cannot open file 'sfg': No such file or directory

#> Error in file(filename, "r", encoding = encoding): cannot open the connection

############# END ###################
## Create PopED database
## (warfarin model for optimization)
#####################################

#for the FO approximation
ind=1

# no occasion defined in this example, so result is zero
LinMatrixL_occ(model_switch=t(poped.db$design$model_switch[ind,,drop=FALSE]),
          xt_ind=t(poped.db$design$xt[ind,,drop=FALSE]),
          x=zeros(0,1),
          a=t(poped.db$design$a[ind,,drop=FALSE]),
          bpop=poped.db$parameters$bpop[,2,drop=FALSE],
          b_ind=zeros(poped.db$parameters$NumRanEff,1),
          bocc_ind=zeros(poped.db$parameters$NumDocc,1),
          iCurrentOcc=1,
          poped.db)["y"]

#> Error in LinMatrixL_occ(model_switch = t(poped.db$design$model_switch[ind,     , drop = FALSE]), xt_ind = t(poped.db$design$xt[ind, , drop = FALSE]),     x = zeros(0, 1), a = t(poped.db$design$a[ind, , drop = FALSE]),     bpop = poped.db$parameters$bpop[, 2, drop = FALSE], b_ind = zeros(poped.db$parameters$NumRanEff,         1), bocc_ind = zeros(poped.db$parameters$NumDocc, 1),     iCurrentOcc = 1, poped.db): object 'poped.db' not found