Header function for optimization routines

Create some output to the screen and a text file that summarizes the problem you are tying to solve.

blockheader(poped.db, name = "Default", iter = NULL,
  e_flag = !(poped.db$settings$d_switch),
  opt_xt = poped.db$settings$optsw[2], opt_a = poped.db$settings$optsw[4],
  opt_x = poped.db$settings$optsw[3],
  opt_samps = poped.db$settings$optsw[1],
  opt_inds = poped.db$settings$optsw[5], fmf = 0, dmf = 0, bpop = NULL,
  d = NULL, docc = NULL, sigma = NULL,
  name_header = poped.db$settings$strOutputFileName,
  file_path = poped.db$settings$strOutputFilePath, out_file = NULL,
  compute_inv = TRUE, trflag = TRUE, header_flag = TRUE, ...)

Arguments

poped.db

A PopED database.

name

The name used for the output file. Combined with name_header and iter. If "" then output is to the screen.

iter

The last number in the name printed to the output file, combined with name.

e_flag

Shuould output be with uncertainty around parameters?

opt_xt

Should the sample times be optimized?

opt_a

Should the continuous design variables be optimized?

opt_x

Should the discrete design variables be optimized?

opt_samps

Are the nuber of sample times per group being optimized?

opt_inds

Are the nuber of individuals per group being optimized?

fmf

The initial value of the FIM. If set to zero then it is computed.

dmf

The inital OFV. If set to zero then it is computed.

bpop

Matrix defining the fixed effects, per row (row number = parameter_number) we should have:

column 1 the type of the distribution for E-family designs (0 = Fixed, 1 = Normal, 2 = Uniform, 3 = User Defined Distribution, 4 = lognormal and 5 = truncated normal)

column 2 defines the mean.

column 3 defines the variance of the distribution (or length of uniform distribution).

Can also just supply the parameter values as a vector c() if no uncertainty around the parameter value is to be used.

d

Matrix defining the diagnonals of the IIV (same logic as for the fixed efects matrix bpop to define uncertainty). One can also just supply the parameter values as a c().

docc

Matrix defining the IOV, the IOV variances and the IOV distribution as for d and bpop.

sigma

Matrix defining the variances can covariances of the residual variability terms of the model. can also just supply the diagnonal parameter values (variances) as a c().

name_header

The initial portion of the file name.

file_path

The path to where the file should be created.

out_file

Which file should the output be directed to? A string, a file handle using file or "" will output to the screen.

compute_inv

should the inverse of the FIM be used to compute expected RSE values? Often not needed except for diagnostic purposes.

trflag

Should the optimization be output to the screen and to a file?

header_flag

Should the header text be printed out?

...

Additional arguments passed to further functions.

Value

fn A file handle (or '' if name='')

See also

Other Helper: blockexp, blockfinal, blockopt

Examples

library(PopED)

############# START #################
## Create PopED database
## (warfarin model for optimization)
#####################################

## Warfarin example from software comparison in:
## Nyberg et al., "Methods and software tools for design evaluation 
##   for population pharmacokinetics-pharmacodynamics studies", 
##   Br. J. Clin. Pharm., 2014. 

## Optimization using an additive + proportional reidual error  
## to avoid sample times at very low concentrations (time 0 or very late samples).

## find the parameters that are needed to define from the structural model
ff.PK.1.comp.oral.sd.CL

#> function(model_switch,xt,parameters,poped.db){
#>   ##-- Model: One comp first order absorption
#>   with(as.list(parameters),{
#>     y=xt
#>     y=(DOSE*Favail*KA/(V*(KA-CL/V)))*(exp(-CL/V*xt)-exp(-KA*xt))
#>     return(list( y= y,poped.db=poped.db))
#>   })
#> }
#> <environment: namespace:PopED>

## -- parameter definition function 
## -- names match parameters in function ff
sfg <- function(x,a,bpop,b,bocc){
  parameters=c(CL=bpop[1]*exp(b[1]),
               V=bpop[2]*exp(b[2]),
               KA=bpop[3]*exp(b[3]),
               Favail=bpop[4],
               DOSE=a[1])
  return(parameters)
}

## -- Define initial design  and design space
poped.db <- create.poped.database(ff_file="ff.PK.1.comp.oral.sd.CL",
                                  fg_file="sfg",
                                  fError_file="feps.add.prop",
                                  bpop=c(CL=0.15, V=8, KA=1.0, Favail=1),
                                  notfixed_bpop=c(1,1,1,0),
                                  d=c(CL=0.07, V=0.02, KA=0.6),
                                  sigma=c(0.01,0.25),
                                  groupsize=32,
                                  xt=c( 0.5,1,2,6,24,36,72,120),
                                  minxt=0.01,
                                  maxxt=120,
                                  a=70,
                                  mina=0.01,
                                  maxa=100)

#> Warning: cannot open file 'sfg': No such file or directory

#> Error in file(filename, "r", encoding = encoding): cannot open the connection

############# END ###################
## Create PopED database
## (warfarin model for optimization)
#####################################

FIM <- evaluate.fim(poped.db)

#> Error in evaluate.fim(poped.db): object 'poped.db' not found

dmf <- det(FIM)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM' not found

blockheader(poped.db,name="")

#> ==============================================================================
#> Optimization of design parameters
#> 

#> Error in blockoptwrt(fn, poped.db$settings$optsw, opt_xt = opt_xt, opt_a = opt_a,     opt_x = opt_x, opt_samps = opt_samps, opt_inds = opt_inds): object 'poped.db' not found

blockheader(name="",iter=1,poped.db)

#> ==============================================================================
#> Optimization of design parameters
#> 

#> Error in blockoptwrt(fn, poped.db$settings$optsw, opt_xt = opt_xt, opt_a = opt_a,     opt_x = opt_x, opt_samps = opt_samps, opt_inds = opt_inds): object 'poped.db' not found

blockheader(name='',
              iter=1,
              poped.db,
              e_flag=FALSE,
              opt_xt=TRUE,
              opt_a=TRUE,opt_x=poped.db$settings$optsw[4],
              opt_samps=poped.db$settings$optsw[1],opt_inds=poped.db$settings$optsw[5],
              fmf=FIM,dmf=dmf,
              bpop=poped.db$parameters$param.pt.val$bpop,
              d=poped.db$parameters$param.pt.val$d,
              docc=poped.db$parameters$docc,sigma=poped.db$parameters$param.pt.val$sigma)

#> Error in blockheader(name = "", iter = 1, poped.db, e_flag = FALSE, opt_xt = TRUE,     opt_a = TRUE, opt_x = poped.db$settings$optsw[4], opt_samps = poped.db$settings$optsw[1],     opt_inds = poped.db$settings$optsw[5], fmf = FIM, dmf = dmf,     bpop = poped.db$parameters$param.pt.val$bpop, d = poped.db$parameters$param.pt.val$d,     docc = poped.db$parameters$docc, sigma = poped.db$parameters$param.pt.val$sigma): object 'dmf' not found

blockheader(name='',
              iter=1,
              poped.db,
              e_flag=TRUE,
              opt_xt=TRUE,
              opt_a=TRUE,opt_x=poped.db$settings$optsw[4],
              opt_samps=poped.db$settings$optsw[1],opt_inds=poped.db$settings$optsw[5],
              fmf=FIM,dmf=dmf,
              bpop=poped.db$parameters$param.pt.val$bpop,
              d=poped.db$parameters$param.pt.val$d,
              docc=poped.db$parameters$docc,sigma=poped.db$parameters$param.pt.val$sigma)

#> Error in blockheader(name = "", iter = 1, poped.db, e_flag = TRUE, opt_xt = TRUE,     opt_a = TRUE, opt_x = poped.db$settings$optsw[4], opt_samps = poped.db$settings$optsw[1],     opt_inds = poped.db$settings$optsw[5], fmf = FIM, dmf = dmf,     bpop = poped.db$parameters$param.pt.val$bpop, d = poped.db$parameters$param.pt.val$d,     docc = poped.db$parameters$docc, sigma = poped.db$parameters$param.pt.val$sigma): object 'dmf' not found

poped.db.1 <- create.poped.database(ff_file="ff.PK.1.comp.oral.sd.CL",
                                  fg_file="sfg",
                                  fError_file="feps.add.prop",
                                  bpop=c(CL=0.15, V=8, KA=1.0, Favail=1),
                                  notfixed_bpop=c(1,1,1,0),
                                  d=c(CL=0.07, V=0.02, KA=0.6),
                                  sigma=c(0.01,0.25),
                                  groupsize=32,
                                  xt=rbind(c( 0.5,1,2,6,24,36,72,120),
                                           c( 0.5,1.1,2,6,24,36,72,120)),
                                  minxt=rbind(c(0,1,1.5,3,20,30,70,118),
                                              c(0.1,1.1,1.6,3.1,20.1,30.1,70.1,118.1)),
                                  maxxt=c(12,13,14,15,26,44,78,120),
                                  a=70,
                                  mina=0,
                                  maxa=100)

#> Warning: cannot open file 'sfg': No such file or directory

#> Error in file(filename, "r", encoding = encoding): cannot open the connection

blockheader(poped.db.1,name="",trflag=2,opt_xt=TRUE)

#> ==============================================================================

#> Error in sprintf(...): object 'poped.db.1' not found