Compute efficiency.

Efficiency calculation between two designs.

efficiency(ofv_init, ofv_final, poped_db, npar = get_fim_size(poped_db),
  ofv_calc_type = poped_db$settings$ofv_calc_type,
  ds_index = poped_db$parameters$ds_index)

Arguments

ofv_init

An initial objective function

ofv_final

A final objective function.

poped_db

a poped database

npar

The number of parameters to use for normalization.

ofv_calc_type

OFV calculation type for FIM

1 = "D-optimality". Determinant of the FIM: det(FIM)

2 = "A-optimality". Inverse of the sum of the expected parameter variances: 1/trace_matrix(inv(FIM))

4 = "lnD-optimality". Natural logarithm of the determinant of the FIM: log(det(FIM))

6 = "Ds-optimality". Ratio of the Determinant of the FIM and the Determinant of the uninteresting rows and columns of the FIM: det(FIM)/det(FIM_u)

7 = Inverse of the sum of the expected parameter RSE: 1/sum(get_rse(FIM,poped.db,use_percent=FALSE))

ds_index

Ds_index is a vector set to 1 if a parameter is uninteresting, otherwise 0. size=(1,num unfixed parameters). First unfixed bpop, then unfixed d, then unfixed docc and last unfixed sigma. Default is the fixed effects being important, everything else not important. Used in conjunction with ofv_calc_type=6.

Value

The specified efficiency value depending on the ofv_calc_type. The attribute "description" tells you how the calculation was made attr(return_vale,"description")

See also

Other FIM: LinMatrixH, LinMatrixLH, LinMatrixL_occ, calc_ofv_and_fim, ed_laplace_ofv, ed_mftot, evaluate.e.ofv.fim, evaluate.fim, gradf_eps, mf3, mf5, mf6, mf7, mf8, mftot0, mftot1, mftot2, mftot3, mftot4, mftot5, mftot6, mftot7, mftot, mf, ofv_criterion, ofv_fim