Structural model: one-compartment, oral absorption, single bolus dose, parameterized using CL.

This is a structural model function that encodes a model that is one-compartment, oral absorption, single bolus dose, parameterized using CL. The function is suitable for input to the create.poped.database function using the ff_file argument.

ff.PK.1.comp.oral.sd.CL(model_switch, xt, parameters, poped.db)

Arguments

model_switch

a vector of values, the same size as xt, identifying which model response should be computed for the corresponding xt value. Used for multiple response models.

xt

a vector of independent variable values (often time).

parameters

A named list of parameter values.

poped.db

a poped database. This can be used to extract information that may be needed in the model file.

Value

A list consisting of:

y the values of the model at the specified points.

poped.db A (potentially modified) poped database.

See also

Other models: feps.add.prop, feps.add, feps.prop, ff.PK.1.comp.oral.md.CL, ff.PK.1.comp.oral.md.KE, ff.PK.1.comp.oral.sd.KE, ff.PKPD.1.comp.oral.md.CL.imax, ff.PKPD.1.comp.sd.CL.emax

Other structural_models: ff.PK.1.comp.oral.md.CL, ff.PK.1.comp.oral.md.KE, ff.PK.1.comp.oral.sd.KE, ff.PKPD.1.comp.oral.md.CL.imax, ff.PKPD.1.comp.sd.CL.emax

Examples

library(PopED)

############# START #################
## Create PopED database
## (warfarin example)
#####################################

## Warfarin example from software comparison in:
## Nyberg et al., "Methods and software tools for design evaluation 
##   for population pharmacokinetics-pharmacodynamics studies", 
##   Br. J. Clin. Pharm., 2014. 

## find the parameters that are needed to define from the structural model
ff.PK.1.comp.oral.sd.CL

#> function(model_switch,xt,parameters,poped.db){
#>   ##-- Model: One comp first order absorption
#>   with(as.list(parameters),{
#>     y=xt
#>     y=(DOSE*Favail*KA/(V*(KA-CL/V)))*(exp(-CL/V*xt)-exp(-KA*xt))
#>     return(list( y= y,poped.db=poped.db))
#>   })
#> }
#> <environment: namespace:PopED>

## -- parameter definition function 
## -- names match parameters in function ff
sfg <- function(x,a,bpop,b,bocc){
  parameters=c(CL=bpop[1]*exp(b[1]),
               V=bpop[2]*exp(b[2]),
               KA=bpop[3]*exp(b[3]),
               Favail=bpop[4],
               DOSE=a[1])
  return(parameters)
}

## -- Define initial design  and design space
poped.db <- create.poped.database(ff_file="ff.PK.1.comp.oral.sd.CL",
                                  fg_file="sfg",
                                  fError_file="feps.prop",
                                  bpop=c(CL=0.15, V=8, KA=1.0, Favail=1),
                                  notfixed_bpop=c(1,1,1,0),
                                  d=c(CL=0.07, V=0.02, KA=0.6),
                                  sigma=0.01,
                                  groupsize=32,
                                  xt=c( 0.5,1,2,6,24,36,72,120),
                                  minxt=0,
                                  maxxt=120,
                                  a=70)

#> Warning: cannot open file 'sfg': No such file or directory

#> Error in file(filename, "r", encoding = encoding): cannot open the connection

############# END ###################
## Create PopED database
## (warfarin example)
#####################################

##  create plot of model without variability 
plot_model_prediction(poped.db)

#> Error in model_prediction(poped.db, model_num_points = model_num_points,     ...): object 'poped.db' not found

## evaluate initial design
FIM <- evaluate.fim(poped.db)

#> Error in evaluate.fim(poped.db): object 'poped.db' not found

FIM

#> Error in eval(expr, envir, enclos): object 'FIM' not found

det(FIM)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM' not found

get_rse(FIM,poped.db)

#> Error in chol(mat): object 'FIM' not found