Model linearization with respect to epsilon.

The function performs a linearization of the model with respect to the residual variability. Derivative of model w.r.t. eps evaluated at eps=0 and b=b_ind.

gradf_eps(model_switch, xt_ind, x, a, bpop, b_ind, bocc_ind, num_eps, poped.db)

Arguments

model_switch
A matrix that is the same size as xt, specifying which model each sample belongs to.

xt_ind
A vector of the individual/group sample times

x
A matrix for the discrete design variables. Each row is a group.

a
A matrix of covariates. Each row is a group.

bpop
The fixed effects parameter values. Supplied as a vector.

b_ind
vector of individual realization of the BSV terms b

bocc_ind
Vector of individual realizations of the BOV terms bocc

num_eps
The number of eps() in the model.

poped.db
A PopED database.

Value

A matrix of size (samples per individual x number of epsilons)

See also

Examples

library(PopED)

############# START #################
## Create PopED database
## (warfarin model for optimization)
#####################################

## Warfarin example from software comparison in:
## Nyberg et al., "Methods and software tools for design evaluation 
##   for population pharmacokinetics-pharmacodynamics studies", 
##   Br. J. Clin. Pharm., 2014. 

## Optimization using an additive + proportional reidual error  
## to avoid sample times at very low concentrations (time 0 or very late samples).

## find the parameters that are needed to define from the structural model
ff.PK.1.comp.oral.sd.CL

#> function (model_switch, xt, parameters, poped.db) 
#> {
#>     with(as.list(parameters), {
#>         y = xt
#>         y = (DOSE * Favail * KA/(V * (KA - CL/V))) * (exp(-CL/V * 
#>             xt) - exp(-KA * xt))
#>         return(list(y = y, poped.db = poped.db))
#>     })
#> }
#> <bytecode: 0x7ffbb9461e08>
#> <environment: namespace:PopED>

## -- parameter definition function 
## -- names match parameters in function ff
sfg <- function(x,a,bpop,b,bocc){
  parameters=c(CL=bpop[1]*exp(b[1]),
               V=bpop[2]*exp(b[2]),
               KA=bpop[3]*exp(b[3]),
               Favail=bpop[4],
               DOSE=a[1])
  return(parameters)
}

## -- Define initial design  and design space
poped.db <- create.poped.database(ff_fun=ff.PK.1.comp.oral.sd.CL,
                                  fg_fun=sfg,
                                  fError_fun=feps.add.prop,
                                  bpop=c(CL=0.15, V=8, KA=1.0, Favail=1),
                                  notfixed_bpop=c(1,1,1,0),
                                  d=c(CL=0.07, V=0.02, KA=0.6),
                                  sigma=c(prop=0.01,add=0.25),
                                  groupsize=32,
                                  xt=c( 0.5,1,2,6,24,36,72,120),
                                  minxt=0.01,
                                  maxxt=120,
                                  a=c(DOSE=70),
                                  mina=c(DOSE=0.01),
                                  maxa=c(DOSE=100))

############# END ###################
## Create PopED database
## (warfarin model for optimization)
#####################################

#for the FO approximation
ind=1
gradf_eps(model_switch=t(poped.db$design$model_switch[ind,,drop=FALSE]),
          xt_ind=t(poped.db$design$xt[ind,,drop=FALSE]),
          x=zeros(0,1),
          a=t(poped.db$design$a[ind,,drop=FALSE]),
          bpop=poped.db$parameters$bpop[,2,drop=FALSE],
          b_ind=zeros(poped.db$parameters$NumRanEff,1),
          bocc_ind=zeros(poped.db$parameters$NumDocc,1),
          num_eps=size(poped.db$parameters$sigma,1),
          poped.db)["dfeps_de0"]

#> $dfeps_de0
#>           [,1] [,2]
#> [1,] 3.4254357    1
#> [2,] 5.4711041    1
#> [3,] 7.3821834    1
#> [4,] 7.9462805    1
#> [5,] 5.6858561    1
#> [6,] 4.5402483    1
#> [7,] 2.3116966    1
#> [8,] 0.9398657    1
#>