Function written to match MATLAB's rand function

Generate random samples from a uniform distribution [0,1] and return in matrix form

rand(dim1, dim2 = NULL)

Arguments

dim1

The dimension of the matrix (if square), otherwise the number of rows.

dim2

The number of colums, if different from the number of rows.

Value

Matrix of random generated samples.

See also

Other MATLAB: cell, diag_matlab, feval, fileparts, isempty, isfield, ones, randn, randperm, size, tic, toc, zeros

Examples

rand(2,3)

#>           [,1]      [,2]      [,3]
#> [1,] 0.8536547 0.4920973 0.4474458
#> [2,] 0.2239750 0.2205318 0.3995766

rand(5)

#>           [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.3102126 0.29990505 0.02671246 0.06518547 0.23379743
#> [2,] 0.1390448 0.25497866 0.76800050 0.78977460 0.87952244
#> [3,] 0.1689644 0.76898410 0.35408669 0.76282817 0.27672004
#> [4,] 0.1637475 0.01289643 0.44717393 0.09521884 0.68885990
#> [5,] 0.9212843 0.72286714 0.10615477 0.21023480 0.08274294