Function written to match MATLAB's randn function

Generate random samples from a standardized normal distribution and return in matrix form.

randn(dim1, dim2 = NULL)

Arguments

dim1

The dimension of the matrix (if square), otherwise the number of rows.

dim2

The number of colums, if different from the number of rows.

Value

Matrix of random generated samples.

See also

Other MATLAB: cell, diag_matlab, feval, fileparts, isempty, isfield, ones, randperm, rand, size, tic, toc, zeros

Examples

randn(2,3)

#>           [,1]       [,2]       [,3]
#> [1,] 1.7658207 -0.2223651 -0.1958017
#> [2,] 0.7747429  1.8902413 -0.5006106

randn(5)

#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] -1.8582041 -1.4957563 -0.4643974 -0.2597059  0.2411745
#> [2,] -0.2910237  1.0476796 -0.8841717  0.9228289  1.2838955
#> [3,]  2.2241945  0.3343878  1.5424870  0.8779387  0.9062239
#> [4,] -1.8195243  1.4035990 -0.2826517 -0.2024283  0.8688575
#> [5,]  1.0301638 -1.2129548 -1.5535861  0.6782348 -2.0846457