Display a summary of output from poped_optim

Display a summary of output from poped_optim

# S3 method for poped_optim
summary(object, ...)

Arguments

object

An object returned from poped_optim to summarize.

...

Additional arguments. Passed to blockfinal.

Examples

library(PopED)

############# START #################
## Create PopED database
## (warfarin model for optimization)
#####################################

## Warfarin example from software comparison in:
## Nyberg et al., "Methods and software tools for design evaluation 
##   for population pharmacokinetics-pharmacodynamics studies", 
##   Br. J. Clin. Pharm., 2014. 

## Optimization using an additive + proportional reidual error  
## to avoid sample times at very low concentrations (time 0 or very late samples).

## find the parameters that are needed to define from the structural model
ff.PK.1.comp.oral.sd.CL

#> function(model_switch,xt,parameters,poped.db){
#>   ##-- Model: One comp first order absorption
#>   with(as.list(parameters),{
#>     y=xt
#>     y=(DOSE*Favail*KA/(V*(KA-CL/V)))*(exp(-CL/V*xt)-exp(-KA*xt))
#>     return(list( y= y,poped.db=poped.db))
#>   })
#> }
#> <environment: namespace:PopED>

## -- parameter definition function 
## -- names match parameters in function ff
sfg <- function(x,a,bpop,b,bocc){
  parameters=c(CL=bpop[1]*exp(b[1]),
               V=bpop[2]*exp(b[2]),
               KA=bpop[3]*exp(b[3]),
               Favail=bpop[4],
               DOSE=a[1])
  return(parameters)
}

## -- Define initial design  and design space
poped.db <- create.poped.database(ff_file="ff.PK.1.comp.oral.sd.CL",
                                  fg_file="sfg",
                                  fError_file="feps.add.prop",
                                  bpop=c(CL=0.15, V=8, KA=1.0, Favail=1),
                                  notfixed_bpop=c(1,1,1,0),
                                  d=c(CL=0.07, V=0.02, KA=0.6),
                                  sigma=c(0.01,0.25),
                                  groupsize=32,
                                  xt=c( 0.5,1,2,6,24,36,72,120),
                                  minxt=0.01,
                                  maxxt=120,
                                  a=70,
                                  mina=0.01,
                                  maxa=100)

############# END ###################
## Create PopED database
## (warfarin model for optimization)
#####################################

##############
# D-family Optimization
##############

# ARS+BFGS+LS optimization of dose
# optimization with just a few iterations
# only to check that things are working
out_1 <- poped_optim(poped.db,opt_a =TRUE,
                      control = list(ARS=list(iter=2),
                                     BFGS=list(maxit=2),
                                     LS=list(line_length=2)),
                      iter_max = 1)

#> ===============================================================================
#> Initial design evaluation
#> 
#> Initial OFV = 55.3964
#> 
#> Initial design expected parameter 
#> relative standard error (%RSE)
#>     Parameter   Values   RSE_0
#>       bpop[1]     0.15    5.10
#>       bpop[2]     8.00    3.03
#>       bpop[3]     1.00   14.26
#>        D[1,1]     0.07   29.76
#>        D[2,2]     0.02   36.68
#>        D[3,3]     0.60   26.75
#>    SIGMA[1,1]     0.01   32.01
#>    SIGMA[2,2]     0.25   25.64
#> 
#> ==============================================================================
#> Optimization of design parameters
#> 
#> * Optimize Covariates
#> 
#> ************* Iteration 1  for all optimization methods***********************
#> 
#> *******************************************
#> Running Adaptive Random Search Optimization
#> *******************************************
#> 
#> Total iterations: 2 
#> Elapsed time: 0.038 seconds.
#> 
#> Final OFV =  55.98047 
#> Parameters: 95.75377 
#> 
#> *******************************************
#> Running L-BFGS-B Optimization
#> *******************************************
#> N = 1, M = 5 machine precision = 2.22045e-16
#> At X0, 0 variables are exactly at the bounds
#> At iterate     0  f=       -55.98  |proj g|=     0.013216
#> 
#> iterations 2
#> function evaluations 3
#> segments explored during Cauchy searches 2
#> BFGS updates skipped 0
#> active bounds at final generalized Cauchy point 1
#> norm of the final projected gradient 0
#> final function value -56.032
#> 
#> F = -56.032
#> final  value -56.032042 
#> stopped after 2 iterations
#> 
#> *******************************************
#> Running Line Search Optimization
#> *******************************************
#> 
#>    Initial parameters: 100 
#>    Initial OFV: 56.03204 
#> 
#>    Searching parameter 1 
#>      No change 
#> 
#>    Elapsed time: 0.059 seconds.
#> 
#>    Final OFV =  56.03204 
#>    Parameters: 100 
#> 
#> *******************************************
#> Stopping criteria testing
#> (Compare between start of iteration and end of iteration)
#> *******************************************
#> Difference in OFV:  0.636
#> Relative difference in OFV:  1.15%
#> Efficiency: 
#>   ((exp(ofv_final) / exp(ofv_init))^(1/n_parameters)) = 1.0827
#> 
#>  Efficiency stopping criteria: 
#>   Is (1.0827 <= 1.001)?   No.
#>   Stopping criteria NOT achieved.
#> 
#> Stopping criteria NOT achieved.
#> 
#> ===============================================================================
#> FINAL RESULTS
#> 
#> Optimized Covariates:
#> Group 1: 100
#> 
#> OFV = 56.032
#> 
#> Efficiency: 
#>   ((exp(ofv_final) / exp(ofv_init))^(1/n_parameters)) = 1.0827
#> 
#> Expected parameter 
#> relative standard error (%RSE):
#>     Parameter   Values   RSE_0     RSE
#>       bpop[1]     0.15    5.10    4.93
#>       bpop[2]     8.00    3.03    2.92
#>       bpop[3]     1.00   14.26   14.11
#>        D[1,1]     0.07   29.76   27.85
#>        D[2,2]     0.02   36.68   34.08
#>        D[3,3]     0.60   26.75   26.32
#>    SIGMA[1,1]     0.01   32.01   23.16
#>    SIGMA[2,2]     0.25   25.64   30.21
#> 
#> Total running time: 0.299 seconds

summary(out_1)

#> ===============================================================================
#> FINAL RESULTS
#> 
#> Optimized Covariates:
#> Group 1: 100
#> 
#> OFV = 56.032
#> 
#> Efficiency: 
#>   ((exp(ofv_final) / exp(ofv_init))^(1/n_parameters)) = 1.0827
#> 
#> Expected parameter 
#> relative standard error (%RSE):
#>     Parameter   Values   RSE_0     RSE
#>       bpop[1]     0.15    5.10    4.93
#>       bpop[2]     8.00    3.03    2.92
#>       bpop[3]     1.00   14.26   14.11
#>        D[1,1]     0.07   29.76   27.85
#>        D[2,2]     0.02   36.68   34.08
#>        D[3,3]     0.60   26.75   26.32
#>    SIGMA[1,1]     0.01   32.01   23.16
#>    SIGMA[2,2]     0.25   25.64   30.21
#> 
#> Total running time: 0.299 seconds