Test if any matrix element is zero.

Function tests is any matrix element is zero. For those elements the function sets those values to the minimum value allowed (not zero). This is used to avoid numerical problems in the FIM calculation.

test_for_zeros(mat, ourzero)

Arguments

mat

A matrix.

ourzero

A matrix the same size as mat with the value that zero should be reassigned to.

Value

A matrix

See also

Other matrix_manipulation: diag_matlab, test_for_max, test_for_min

Examples

test_for_zeros(cbind(2,3,0,5,6),1e-18)

#>      [,1] [,2]  [,3] [,4] [,5]
#> [1,]    2    3 1e-18    5    6

test_for_zeros(zeros(6),1e-7)

#>       [,1]  [,2]  [,3]  [,4]  [,5]  [,6]
#> [1,] 1e-07 1e-07 1e-07 1e-07 1e-07 1e-07
#> [2,] 1e-07 1e-07 1e-07 1e-07 1e-07 1e-07
#> [3,] 1e-07 1e-07 1e-07 1e-07 1e-07 1e-07
#> [4,] 1e-07 1e-07 1e-07 1e-07 1e-07 1e-07
#> [5,] 1e-07 1e-07 1e-07 1e-07 1e-07 1e-07
#> [6,] 1e-07 1e-07 1e-07 1e-07 1e-07 1e-07

test_for_zeros(cbind(1,2,3,NA,0),1e-5)

#>      [,1] [,2] [,3] [,4]  [,5]
#> [1,]    1    2    3   NA 1e-05