Timer function (as in MATLAB)

Funtion to stop a timer. Start with tic().

toc(echo = TRUE, name = ".poped_savedTime")

Arguments

echo

Print time to screen?

name

The saved name of the time object.

Note

This is a modified version of the same function in the matlab R-package.

See also

Other MATLAB: cell, diag_matlab, feval, fileparts, isempty, isfield, ones, randn, randperm, rand, size, tic, zeros

Examples

tic()
toc()

#> Elapsed time: 0.002 seconds.

tic(name="foo")
toc()

#> Elapsed time: 0.004 seconds.

tic()
toc()

#> Elapsed time: 0.001 seconds.

toc()

#> Elapsed time: 0.002 seconds.

tic()
toc(name="foo")

#> Elapsed time: 0.006 seconds.