Compute the autofocus portion of the stochastic gradient routine

Compute the autofocus portion of the stochastic gradient routine

calc_autofocus(m, ni_var, dmf, varopt, varopto, maxvar, minvar, gradvar,
  normgvar, avar, model_switch, groupsize, xtopt, xopt, aopt, ni, bpop, d,
  sigma, docc, poped.db)

Arguments

m

Number of groups in the study. Each individual in a group will have the same design.

ni_var

The ni_var.

dmf

The inital OFV. If set to zero then it is computed.

varopt

The varopt.

varopto

The varopto.

maxvar

The maxvar.

minvar

The minvar.

gradvar

The gradvar.

normgvar

The normgvar.

avar

The avar.

model_switch

A matrix that is the same size as xt, specifying which model each sample belongs to.

groupsize

A vector of the numer of individuals in each group.

xtopt

The optimal sampling times matrix.

xopt

The optimal discrete design variables matrix.

aopt

The optimal continuous design variables matrix.

ni

A vector of the number of samples in each group.

bpop

Matrix defining the fixed effects, per row (row number = parameter_number) we should have:

column 1 the type of the distribution for E-family designs (0 = Fixed, 1 = Normal, 2 = Uniform, 3 = User Defined Distribution, 4 = lognormal and 5 = truncated normal)

column 2 defines the mean.

column 3 defines the variance of the distribution (or length of uniform distribution).

Can also just supply the parameter values as a vector c() if no uncertainty around the parameter value is to be used.

d

Matrix defining the diagnonals of the IIV (same logic as for the fixed efects matrix bpop to define uncertainty). One can also just supply the parameter values as a c().

sigma

Matrix defining the variances can covariances of the residual variability terms of the model. can also just supply the diagnonal parameter values (variances) as a c().

docc

Matrix defining the IOV, the IOV variances and the IOV distribution as for d and bpop.

poped.db

A PopED database.

Value

A list containing:

See also

Other Optimize: Doptim, LEDoptim, RS_opt_gen, RS_opt, a_line_search, bfgsb_min, calc_ofv_and_grad, mfea, optim_ARS, optim_LS, poped_optimize, poped_optim

Examples

library(PopED)

############# START #################
## Create PopED database
## (warfarin model for optimization)
#####################################

## Warfarin example from software comparison in:
## Nyberg et al., "Methods and software tools for design evaluation 
##   for population pharmacokinetics-pharmacodynamics studies", 
##   Br. J. Clin. Pharm., 2014. 

## Optimization using an additive + proportional reidual error  
## to avoid sample times at very low concentrations (time 0 or very late samples).

## find the parameters that are needed to define from the structural model
ff.PK.1.comp.oral.sd.CL

#> function(model_switch,xt,parameters,poped.db){
#>   ##-- Model: One comp first order absorption
#>   with(as.list(parameters),{
#>     y=xt
#>     y=(DOSE*Favail*KA/(V*(KA-CL/V)))*(exp(-CL/V*xt)-exp(-KA*xt))
#>     return(list( y= y,poped.db=poped.db))
#>   })
#> }
#> <environment: namespace:PopED>

## -- parameter definition function 
## -- names match parameters in function ff
sfg <- function(x,a,bpop,b,bocc){
  parameters=c(CL=bpop[1]*exp(b[1]),
               V=bpop[2]*exp(b[2]),
               KA=bpop[3]*exp(b[3]),
               Favail=bpop[4],
               DOSE=a[1])
  return(parameters)
}

## -- Define initial design  and design space
poped.db <- create.poped.database(ff_file="ff.PK.1.comp.oral.sd.CL",
                                  fg_file="sfg",
                                  fError_file="feps.add.prop",
                                  bpop=c(CL=0.15, V=8, KA=1.0, Favail=1),
                                  notfixed_bpop=c(1,1,1,0),
                                  d=c(CL=0.07, V=0.02, KA=0.6),
                                  sigma=c(0.01,0.25),
                                  groupsize=32,
                                  xt=c( 0.5,1,2,6,24,36,72,120),
                                  minxt=0.01,
                                  maxxt=120,
                                  a=70,
                                  mina=0.01,
                                  maxa=100)

#> Warning: cannot open file 'sfg': No such file or directory

#> Error in file(filename, "r", encoding = encoding): cannot open the connection

############# END ###################
## Create PopED database
## (warfarin model for optimization)
#####################################

## Not run: ------------------------------------
#   
# # Stochastic gradient search, DOSE and sample time optimization
# sg.output <- poped_optimize(poped.db,opt_xt=1,opt_a=1, 
#                             bUseRandomSearch= 0,
#                             bUseStochasticGradient = 1,
#                             bUseBFGSMinimizer = 0,
#                             bUseLineSearch = 0,
#                             sgit=20)
# 
## ---------------------------------------------