Evaluate the Fisher Information Matrix (FIM)

Compute the FIM given the model, parameters, design and methods defined in the PopED database. Some of the arguments coming from the PopED database can be overwritten; by default these arguments are NULL in the function, if they are supplied then they are used instead of the arguments from the PopED database.

evaluate.fim(poped.db, fim.calc.type = NULL, approx.method = NULL,
  FOCE.num = NULL, bpop.val = NULL, d_full = NULL, docc_full = NULL,
  sigma_full = NULL, model_switch = NULL, ni = NULL, xt = NULL,
  x = NULL, a = NULL, groupsize = NULL, deriv.type = NULL, ...)

Arguments

poped.db

A PopED database.

fim.calc.type

The method used for calculating the FIM. Potential values:

0 = Full FIM. No assumption that fixed and random effects are uncorrelated. See mftot0.

1 = Reduced FIM. Assume that there is no correlation in the FIM between the fixed and random effects, and set these elements in the FIM to zero. See mftot1.

2 = weighted models (placeholder).

3 = Not currently used.

4 = Reduced FIM and computing all derivatives with respect to the standard deviation of the residual unexplained variation (sqrt(SIGMA) in NONMEM). This matches what is done in PFIM, and assumes that the standard deviation of the residual unexplained variation is the estimated parameter (NOTE: NONMEM estimates the variance of the resudual unexplained variation by default). See mftot4.

5 = Full FIM parameterized with A,B,C matrices & derivative of variance. See mftot5.

6 = Calculate one model switch at a time, good for large matrices. See mftot6.

7 = Reduced FIM parameterized with A,B,C matrices & derivative of variance See mftot7.

approx.method

Approximation method for model, 0=FO, 1=FOCE, 2=FOCEI, 3=FOI

FOCE.num

Number indivduals in each step of FOCE approximation method

bpop.val

The fixed effects parameter values. Supplied as a vector.

d_full

A between subject variability matrix (OMEGA in NONMEM).

docc_full

A between occasion variability matrix.

sigma_full

A residual unexplained variability matrix (SIGMA in NONMEM).

model_switch

A matrix that is the same size as xt, specifying which model each sample belongs to.

ni

A vector of the number of samples in each group.

xt

A matrix of sample times. Each row is a vector of sample times for a group.

x

A matrix for the discrete design variables. Each row is a group.

a

A matrix of covariates. Each row is a group.

groupsize

A vector of the numer of individuals in each group.

deriv.type

A number indicating the type of derivative to use:

0=Complex difference

1=Central difference

20=Analytic derivative (placeholder)

30=Automatic differentiation (placeholder)

...

Other arguments passed to the function.

Value

The FIM.

See also

Other evaluate_design: evaluate_design, get_rse, model_prediction, plot_efficiency_of_windows, plot_model_prediction

Other evaluate_FIM: calc_ofv_and_fim, evaluate.e.ofv.fim, ofv_fim

Other FIM: LinMatrixH, LinMatrixLH, LinMatrixL_occ, calc_ofv_and_fim, ed_laplace_ofv, ed_mftot, efficiency, evaluate.e.ofv.fim, gradf_eps, mf3, mf5, mf6, mf7, mf8, mftot0, mftot1, mftot2, mftot3, mftot4, mftot5, mftot6, mftot7, mftot, mf, ofv_criterion, ofv_fim

Examples

## Warfarin example from software comparison in:
## Nyberg et al., "Methods and software tools for design evaluation 
##   for population pharmacokinetics-pharmacodynamics studies", 
##   Br. J. Clin. Pharm., 2014. 

library(PopED)

## find the parameters that are needed to define from the structural model
ff.PK.1.comp.oral.md.CL

#> function(model_switch,xt,parameters,poped.db){
#>   ##-- Model: One comp first order absorption
#>   ## -- Analytic solution for both mutiple and single dosing
#>   with(as.list(parameters),{
#>     y=xt
#>     N = floor(xt/TAU)+1
#>     y=(DOSE*Favail/V)*(KA/(KA - CL/V)) * 
#>       (exp(-CL/V * (xt - (N - 1) * TAU)) * (1 - exp(-N * CL/V * TAU))/(1 - exp(-CL/V * TAU)) - 
#>          exp(-KA * (xt - (N - 1) * TAU)) * (1 - exp(-N * KA * TAU))/(1 - exp(-KA * TAU)))  
#>     return(list( y=y,poped.db=poped.db))
#>   })
#> }
#> <environment: namespace:PopED>

## -- parameter definition function 
## -- names match parameters in function ff
sfg <- function(x,a,bpop,b,bocc){
  parameters=c(CL=bpop[1]*exp(b[1]),
               V=bpop[2]*exp(b[2]),
               KA=bpop[3]*exp(b[3]),
               Favail=bpop[4],
               DOSE=a[1])
    return(parameters)
}

## -- Define initial design  and design space
poped.db <- create.poped.database(ff_file="ff.PK.1.comp.oral.sd.CL",
                                  fg_file="sfg",
                                  fError_file="feps.prop",
                                  bpop=c(CL=0.15, V=8, KA=1.0, Favail=1),
                                  notfixed_bpop=c(1,1,1,0),
                                  d=c(CL=0.07, V=0.02, KA=0.6),
                                  sigma=0.01,
                                  groupsize=32,
                                  xt=c( 0.5,1,2,6,24,36,72,120),
                                  minxt=0,
                                  maxxt=120,
                                  a=70)

#> Warning: cannot open file 'sfg': No such file or directory

#> Error in file(filename, "r", encoding = encoding): cannot open the connection

## evaluate initial design with the reduced FIM
FIM.1 <- evaluate.fim(poped.db)

#> Error in evaluate.fim(poped.db): object 'poped.db' not found

FIM.1

#> Error in eval(expr, envir, enclos): object 'FIM.1' not found

det(FIM.1)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM.1' not found

get_rse(FIM.1,poped.db)

#> Error in chol(mat): object 'FIM.1' not found

## evaluate initial design with the full FIM
FIM.0 <- evaluate.fim(poped.db,fim.calc.type=0)

#> Error in evaluate.fim(poped.db, fim.calc.type = 0): object 'poped.db' not found

FIM.0

#> Error in eval(expr, envir, enclos): object 'FIM.0' not found

det(FIM.0)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM.0' not found

get_rse(FIM.0,poped.db,fim.calc.type=0)

#> Error in poped.db$settings$iFIMCalculationType <- fim.calc.type: object 'poped.db' not found

## evaluate initial design with the reduced FIM 
## computing all derivatives with respect to the 
## standard deviation of the residual unexplained variation 
FIM.4 <- evaluate.fim(poped.db,fim.calc.type=4)

#> Error in evaluate.fim(poped.db, fim.calc.type = 4): object 'poped.db' not found

FIM.4

#> Error in eval(expr, envir, enclos): object 'FIM.4' not found

det(FIM.4)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM.4' not found

get_rse(FIM.4,poped.db,fim.calc.type=4)

#> Error in poped.db$settings$iFIMCalculationType <- fim.calc.type: object 'poped.db' not found

## evaluate initial design with the full FIM with A,B,C matricies
## should give same answer as fim.calc.type=0
FIM.5 <- evaluate.fim(poped.db,fim.calc.type=5)

#> Error in evaluate.fim(poped.db, fim.calc.type = 5): object 'poped.db' not found

FIM.5

#> Error in eval(expr, envir, enclos): object 'FIM.5' not found

det(FIM.5)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM.5' not found

get_rse(FIM.5,poped.db,fim.calc.type=5)

#> Error in poped.db$settings$iFIMCalculationType <- fim.calc.type: object 'poped.db' not found

## evaluate initial design with the reduced FIM with 
## A,B,C matricies and derivative of variance
## should give same answer as fim.calc.type=1 (default)
FIM.7 <- evaluate.fim(poped.db,fim.calc.type=7)

#> Error in evaluate.fim(poped.db, fim.calc.type = 7): object 'poped.db' not found

FIM.7

#> Error in eval(expr, envir, enclos): object 'FIM.7' not found

det(FIM.7)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM.7' not found

get_rse(FIM.7,poped.db,fim.calc.type=7)

#> Error in poped.db$settings$iFIMCalculationType <- fim.calc.type: object 'poped.db' not found