Compute the expected parameter relative standard errors

This function computes the expected relative standard errors of a model given a design and a previously computed FIM.

get_rse(fim, poped.db, bpop = poped.db$parameters$bpop[, 2],
  d = poped.db$parameters$d[, 2], docc = poped.db$parameters$docc,
  sigma = poped.db$parameters$sigma, use_percent = TRUE,
  prior_fim = poped.db$settings$prior_fim, ...)

Arguments

fim
A Fisher Information Matrix (FIM).

poped.db
A PopED database.

bpop
A vector containing the values of the fixed effects used to compute the fim.

d
A vector containing the values of the diagonals of the between subject variability matrix.

docc
Matrix defining the IOV, the IOV variances and the IOV distribution as for d and bpop.

sigma
Matrix defining the variances can covariances of the residual variability terms of the model. can also just supply the diagonal parameter values (variances) as a c().

use_percent
Should RSE be reported as percent?

prior_fim
A prior FIM to be added to the fim. Should be the same size as the fim.

...
Additional arguments passed to inv.

Value

A named list of RSE values. If the estimated parameter is assumed to be zero then for that parameter the standard error is returned.

See also

Examples

## Warfarin example from software comparison in:
## Nyberg et al., "Methods and software tools for design evaluation 
##   for population pharmacokinetics-pharmacodynamics studies", 
##   Br. J. Clin. Pharm., 2014. 

library(PopED)

## find the parameters that are needed to define from the structural model
ff.PK.1.comp.oral.md.CL

#> function(model_switch,xt,parameters,poped.db){
#>   ##-- Model: One comp first order absorption
#>   ## -- Analytic solution for both mutiple and single dosing
#>   with(as.list(parameters),{
#>     y=xt
#>     N = floor(xt/TAU)+1
#>     y=(DOSE*Favail/V)*(KA/(KA - CL/V)) * 
#>       (exp(-CL/V * (xt - (N - 1) * TAU)) * (1 - exp(-N * CL/V * TAU))/(1 - exp(-CL/V * TAU)) - 
#>          exp(-KA * (xt - (N - 1) * TAU)) * (1 - exp(-N * KA * TAU))/(1 - exp(-KA * TAU)))  
#>     return(list( y=y,poped.db=poped.db))
#>   })
#> }
#> <environment: namespace:PopED>

## -- parameter definition function 
## -- names match parameters in function ff
sfg <- function(x,a,bpop,b,bocc){
  parameters=c(CL=bpop[1]*exp(b[1]),
               V=bpop[2]*exp(b[2]),
               KA=bpop[3]*exp(b[3]),
               Favail=bpop[4],
               DOSE=a[1])
    return(parameters)
}

## -- Define initial design  and design space
poped.db <- create.poped.database(ff_file="ff.PK.1.comp.oral.sd.CL",
                                  fg_file="sfg",
                                  fError_file="feps.prop",
                                  bpop=c(CL=0.15, V=8, KA=1.0, Favail=1),
                                  notfixed_bpop=c(1,1,1,0),
                                  d=c(CL=0.07, V=0.02, KA=0.6),
                                  sigma=0.01,
                                  groupsize=32,
                                  xt=c( 0.5,1,2,6,24,36,72,120),
                                  minxt=0,
                                  maxxt=120,
                                  a=70)

#> Warning: cannot open file 'sfg': No such file or directory

#> Error in file(filename, "r", encoding = encoding): cannot open the connection

## evaluate initial design with the reduced FIM
FIM.1 <- evaluate.fim(poped.db)

#> Error in evaluate.fim(poped.db): object 'poped.db' not found

FIM.1

#> Error in eval(expr, envir, enclos): object 'FIM.1' not found

det(FIM.1)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM.1' not found

get_rse(FIM.1,poped.db)

#> Error in size(...): object 'poped.db' not found

## evaluate initial design with the full FIM
FIM.0 <- evaluate.fim(poped.db,fim.calc.type=0)

#> Error in evaluate.fim(poped.db, fim.calc.type = 0): object 'poped.db' not found

FIM.0

#> Error in eval(expr, envir, enclos): object 'FIM.0' not found

det(FIM.0)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM.0' not found

get_rse(FIM.0,poped.db,fim.calc.type=0)

#> Error in size(...): object 'poped.db' not found

## evaluate initial design with the reduced FIM 
## computing all derivatives with respect to the 
## standard deviation of the residual unexplained variation 
FIM.4 <- evaluate.fim(poped.db,fim.calc.type=4)

#> Error in evaluate.fim(poped.db, fim.calc.type = 4): object 'poped.db' not found

FIM.4

#> Error in eval(expr, envir, enclos): object 'FIM.4' not found

det(FIM.4)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM.4' not found

get_rse(FIM.4,poped.db,fim.calc.type=4)

#> Error in size(...): object 'poped.db' not found

## evaluate initial design with the full FIM with A,B,C matricies
## should give same answer as fim.calc.type=0
FIM.5 <- evaluate.fim(poped.db,fim.calc.type=5)

#> Error in evaluate.fim(poped.db, fim.calc.type = 5): object 'poped.db' not found

FIM.5

#> Error in eval(expr, envir, enclos): object 'FIM.5' not found

det(FIM.5)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM.5' not found

get_rse(FIM.5,poped.db,fim.calc.type=5)

#> Error in size(...): object 'poped.db' not found

## evaluate initial design with the reduced FIM with 
## A,B,C matricies and derivative of variance
## should give same answer as fim.calc.type=1 (default)
FIM.7 <- evaluate.fim(poped.db,fim.calc.type=7)

#> Error in evaluate.fim(poped.db, fim.calc.type = 7): object 'poped.db' not found

FIM.7

#> Error in eval(expr, envir, enclos): object 'FIM.7' not found

det(FIM.7)

#> Error in determinant(x, logarithm = TRUE, ...): object 'FIM.7' not found

get_rse(FIM.7,poped.db,fim.calc.type=7)

#> Error in size(...): object 'poped.db' not found

## evaluate FIM and rse with prior FIM.1
poped.db.prior = create.poped.database(poped.db, prior_fim = FIM.1)

#> Error in poped.choose(popedInput$settings$poped_version, packageVersion("PopED")): object 'poped.db' not found

FIM.1.prior <- evaluate.fim(poped.db.prior)

#> Error in evaluate.fim(poped.db.prior): object 'poped.db.prior' not found

all.equal(FIM.1.prior,FIM.1) # the FIM is only computed from the design in the poped.db

#> Error in all.equal(FIM.1.prior, FIM.1): object 'FIM.1.prior' not found

get_rse(FIM.1.prior,poped.db.prior) # the RSE is computed with the prior information

#> Error in size(...): object 'poped.db.prior' not found