Optimization Using Adaptive Random Search.

Optimize an objective function using an adaptive random search algorithm. The function works for both discrete and continuous optimization parameters and allows for box-constraints and sets of allowed values.

optim_ARS(par, fn, lower = NULL, upper = NULL, allowed_values = NULL,
  loc_fac = 4, no_bounds_sd = par, iter = 400, iter_adapt = 50,
  adapt_scale = 1, max_run = 200, trace = TRUE, trace_iter = 5,
  new_par_max_it = 200, maximize = F, parallel = F,
  parallel_type = NULL, num_cores = NULL, seed = round(runif(1, 0,
  1e+07)), allow_replicates = TRUE, generator = NULL, ...)

Arguments

par

A vector of initial values for the parameters to be optimized over.

fn

A function to be minimized (or maximized), with first argument the vector of parameters over which minimization is to take place. It should return a scalar result.

lower

Lower bounds on the parameters. A vector.

upper

Upper bounds on the parameters. A vector.

allowed_values

A list containing allowed values for each parameter list(par1=c(2,3,4,5,6),par2=c(5,6,7,8)). A vector containing allowed values for all parameters is also allowed c(2,3,4,5,6).

loc_fac

Locality factor for determining the standard deviation of the sampling distribution around the current position of the parameters. The initial standard deviation is normally calculated as (upper - lower)/loc_fac except in cases when there are no upper or lower limits (e.g. when upper=Inf or lower=-Inf).

no_bounds_sd

The standard deviation of the sampling distribution around the current position of the parameters when there are no upper or lower limits (e.g. when upper=Inf or lower=-Inf).

iter

The number of iterations for the algorithm to perfrom (this is a maximum number, it could be less).

iter_adapt

The number of iterations before adapting (shrinking) the parameter search space.

adapt_scale

The scale for adapting the size of the sampling distribution. The adaptation of the standard deviation of the sampling distribution around the current position of the parameters is done after iter_adapt iteration with no change in the best objective function. When adapting, the standard deviation of the sampling distribution is calculated as (upper - lower)/(loc_fac*ff*adapt_scale) where ff starts at 1 and increases by 1 for each adaptation.

max_run

The maximum number of iterations to run without a change in the best parameter estimates.

trace

Should the algorithm output results intermittently.

trace_iter

How many interations between each update to the screen about the result of the search.

new_par_max_it

The algorithm randomly chooses samples based on the current best set of parameters. If when drawing these samples the new parameter set has already been tested then a new draw is performed. After new_par_max_it draws, with no new parameter sets, then the algorithm stops.

maximize

Should the function be maximized? Default is to minimize.

parallel

Should we use parallelize the computation.

parallel_type

Which type of parallelization should be used? Can be "snow" or "multicore". "snow" works on Linux-like systems & Windows. "multicore" works only on Linux-like systems. By default this is chosen for you depending on your operating system. See start_parallel.

num_cores

The number of cores to use in the parallelization. By default is set to the number output from parallel::detectCores(). See start_parallel.

seed

The random seed to use in the algorithm,

allow_replicates

Should the algorithm allow parameters to have the same value?

generator

A user-defined function that generates new parameter sets to try in the algorithm. See examples below.

...

Additional arguments passed to fn and start_parallel.

References

M. Foracchia, A.C. Hooker, P. Vicini and A. Ruggeri, "PopED, a software fir optimal experimental design in population kinetics", Computer Methods and Programs in Biomedicine, 74, 2004.

J. Nyberg, S. Ueckert, E.A. Stroemberg, S. Hennig, M.O. Karlsson and A.C. Hooker, "PopED: An extended, parallelized, nonlinear mixed effects models optimal design tool", Computer Methods and Programs in Biomedicine, 108, 2012.

See also

Other Optimize: Doptim, LEDoptim, RS_opt_gen, RS_opt, a_line_search, bfgsb_min, calc_autofocus, calc_ofv_and_grad, mfea, optim_LS, poped_optimize, poped_optim

Examples

## "wild" function , global minimum at about -15.81515
fw <- function(x) 10*sin(0.3*x)*sin(1.3*x^2) + 0.00001*x^4 + 0.2*x+80

# optimization with fewer function evaluations compared to SANN
res1 <- optim_ARS(50, fw,lower = -50, upper=100)

#> It.   5 | OFV = 132.9
#> It.  10 | OFV = 73.5237
#> It.  15 | OFV = 73.5237
#> It.  20 | OFV = 68.3619
#> It.  25 | OFV = 68.3619
#> It.  30 | OFV = 68.3619
#> It.  35 | OFV = 68.3619
#> It.  40 | OFV = 68.3619
#> It.  45 | OFV = 68.3619
#> It.  50 | OFV = 68.3619
#> It.  55 | OFV = 68.3619
#> It.  60 | OFV = 68.3619
#> It.  65 | OFV = 68.3619
#> It.  70 | OFV = 68.3619
#> It.  75 | OFV = 68.3619
#> It.  80 | OFV = 68.3619
#> It.  85 | OFV = 68.3619
#> It.  90 | OFV = 68.3619
#> It.  95 | OFV = 68.3619
#> It. 100 | OFV = 68.0983
#> It. 105 | OFV = 68.0983
#> It. 110 | OFV = 68.0983
#> It. 115 | OFV = 68.0983
#> It. 120 | OFV = 68.0983
#> It. 125 | OFV = 68.0983
#> It. 130 | OFV = 68.0983
#> It. 135 | OFV = 68.0983
#> It. 140 | OFV = 68.0983
#> It. 145 | OFV = 68.0983
#> It. 150 | OFV = 68.0983
#> It. 155 | OFV = 68.0983
#> It. 160 | OFV = 68.0983
#> It. 165 | OFV = 68.0983
#> It. 170 | OFV = 68.0983
#> It. 175 | OFV = 68.0983
#> It. 180 | OFV = 68.0983
#> It. 185 | OFV = 68.0128
#> It. 190 | OFV = 68.0128
#> It. 195 | OFV = 67.966
#> It. 200 | OFV = 67.966
#> It. 205 | OFV = 67.966
#> It. 210 | OFV = 67.966
#> It. 215 | OFV = 67.966
#> It. 220 | OFV = 67.966
#> It. 225 | OFV = 67.966
#> It. 230 | OFV = 67.966
#> It. 235 | OFV = 67.966
#> It. 240 | OFV = 67.966
#> It. 245 | OFV = 67.966
#> It. 250 | OFV = 67.966
#> It. 255 | OFV = 67.966
#> It. 260 | OFV = 67.966
#> It. 265 | OFV = 67.966
#> It. 270 | OFV = 67.966
#> It. 275 | OFV = 67.966
#> It. 280 | OFV = 67.966
#> It. 285 | OFV = 67.966
#> It. 290 | OFV = 67.966
#> It. 295 | OFV = 67.966
#> It. 300 | OFV = 67.966
#> It. 305 | OFV = 67.966
#> It. 310 | OFV = 67.966
#> It. 315 | OFV = 67.966
#> It. 320 | OFV = 67.966
#> It. 325 | OFV = 67.966
#> It. 330 | OFV = 67.966
#> It. 335 | OFV = 67.966
#> It. 340 | OFV = 67.966
#> It. 345 | OFV = 67.966
#> It. 350 | OFV = 67.966
#> It. 355 | OFV = 67.966
#> It. 360 | OFV = 67.966
#> It. 365 | OFV = 67.966
#> It. 370 | OFV = 67.966
#> It. 375 | OFV = 67.966
#> It. 380 | OFV = 67.966
#> It. 385 | OFV = 67.966
#> It. 390 | OFV = 67.966
#> Maximum number of identical optimal values reached (max_run=200), optimization stopped.
#> 
#> Total iterations: 390 
#> Elapsed time: 0.252 seconds.
#> 
#> Final OFV =  67.96598 
#> Parameters: -15.97471 
#> 

# often not as good performance when upper and lower bounds are poor
res2 <- optim_ARS(50, fw, lower=-Inf,upper=Inf)

#> It.   5 | OFV = 81.9565
#> It.  10 | OFV = 73.7888
#> It.  15 | OFV = 73.1635
#> It.  20 | OFV = 73.1635
#> It.  25 | OFV = 73.1635
#> It.  30 | OFV = 73.1635
#> It.  35 | OFV = 73.1635
#> It.  40 | OFV = 73.1635
#> It.  45 | OFV = 72.3227
#> It.  50 | OFV = 70.3111
#> It.  55 | OFV = 70.3111
#> It.  60 | OFV = 67.5328
#> It.  65 | OFV = 67.5328
#> It.  70 | OFV = 67.5328
#> It.  75 | OFV = 67.5328
#> It.  80 | OFV = 67.5328
#> It.  85 | OFV = 67.5328
#> It.  90 | OFV = 67.5328
#> It.  95 | OFV = 67.5328
#> It. 100 | OFV = 67.5328
#> It. 105 | OFV = 67.5328
#> It. 110 | OFV = 67.5328
#> It. 115 | OFV = 67.5328
#> It. 120 | OFV = 67.5328
#> It. 125 | OFV = 67.5328
#> It. 130 | OFV = 67.5328
#> It. 135 | OFV = 67.5328
#> It. 140 | OFV = 67.5328
#> It. 145 | OFV = 67.5328
#> It. 150 | OFV = 67.5328
#> It. 155 | OFV = 67.5328
#> It. 160 | OFV = 67.5328
#> It. 165 | OFV = 67.5328
#> It. 170 | OFV = 67.5328
#> It. 175 | OFV = 67.5328
#> It. 180 | OFV = 67.5328
#> It. 185 | OFV = 67.5328
#> It. 190 | OFV = 67.5328
#> It. 195 | OFV = 67.5328
#> It. 200 | OFV = 67.5328
#> It. 205 | OFV = 67.5328
#> It. 210 | OFV = 67.5328
#> It. 215 | OFV = 67.5328
#> It. 220 | OFV = 67.5328
#> It. 225 | OFV = 67.5328
#> It. 230 | OFV = 67.5328
#> It. 235 | OFV = 67.5328
#> It. 240 | OFV = 67.5328
#> It. 245 | OFV = 67.5328
#> It. 250 | OFV = 67.5328
#> It. 255 | OFV = 67.5328
#> Maximum number of identical optimal values reached (max_run=200), optimization stopped.
#> 
#> Total iterations: 256 
#> Elapsed time: 0.125 seconds.
#> 
#> Final OFV =  67.53284 
#> Parameters: -16.11868 
#> 

# Only integer values allowed
res_int <- optim_ARS(50, fw, allowed_values = seq(-50,100,by=1))

#> It.   5 | OFV = 88.1321
#> It.  10 | OFV = 74.3212
#> It.  15 | OFV = 70.2605
#> It.  20 | OFV = 70.2605
#> It.  25 | OFV = 70.2605
#> It.  30 | OFV = 70.2605
#> It.  35 | OFV = 70.2605
#> It.  40 | OFV = 70.1897
#> It.  45 | OFV = 70.1897
#> It.  50 | OFV = 70.1897
#> It.  55 | OFV = 70.1897
#> It.  60 | OFV = 70.1897
#> It.  65 | OFV = 70.1897
#> It.  70 | OFV = 70.1897
#> It.  75 | OFV = 70.1897
#> It.  80 | OFV = 70.1897
#> It.  85 | OFV = 70.1897
#> It.  90 | OFV = 70.1897
#> It.  95 | OFV = 68.5368
#> It. 100 | OFV = 68.5368
#> It. 105 | OFV = 68.5368
#> It. 110 | OFV = 68.5368
#> It. 115 | OFV = 68.5368
#> It. 120 | OFV = 68.5368
#> It. 125 | OFV = 68.5368
#> It. 130 | OFV = 68.5368
#> It. 135 | OFV = 68.5368
#> It. 140 | OFV = 68.5368
#> It. 145 | OFV = 68.5368
#> It. 150 | OFV = 68.5368
#> Maximum number of duplicate parameter samples reached (new_par_max_it=200), optimization stopped.
#> 
#> Total iterations: 152 
#> Elapsed time: 0.388 seconds.
#> 
#> Final OFV =  68.53679 
#> Parameters: -17 
#> 

## Not run: ------------------------------------
#  
#   #plot of the function and solutions
#   require(graphics)
#   plot(fw, -50, 50, n = 1000, main = "Minimizing 'wild function'")
#   points(-15.81515, fw(-15.81515), pch = 16, col = "red", cex = 1)
#   points(res1$par, res1$ofv, pch = 16, col = "green", cex = 1)
#   points(res2$par, res2$ofv, pch = 16, col = "blue", cex = 1)
## --------------------------------------------- 

# optim_ARS does not work great for hard to find minima on flat surface:
# Rosenbrock Banana function
# f(x, y) = (a-x)^2 + b(y-x^2)^2
# global minimum at (x, y)=(a, a^2), where f(x, y)=0. 
# Usually a = 1 and b = 100.
## Not run: ------------------------------------
#  
#   fr <- function(x,a=1,b=100) {   
#     x1 <- x[1]
#     x2 <- x[2]
#     b*(x2 - x1*x1)^2 + (a - x1)^2
#   }
#   
#   res3 <- optim_ARS(c(-1.2,1), fr,lower = -5, upper = 5)
#   
#   # plot the surface
#   x <- seq(-50, 50, length= 30)
#   y <- x
#   f <- function(x,y){apply(cbind(x,y),1,fr)}
#   z <- outer(x, y, f)
#   persp(x, y, z, theta = 30, phi = 30, expand = 0.5, col = "lightblue", ticktype="detailed") -> res
#   points(trans3d(1, 1, 0, pmat = res), col = 2, pch = 16,cex=2)
#   points(trans3d(res3$par[1], res3$par[1], res3$ofv, pmat = res), col = "green", pch = 16,cex=2)
## ---------------------------------------------

# box constraints
flb <- function(x){
  p <- length(x)
  sum(c(1, rep(4, p-1)) * (x - c(1, x[-p])^2)^2)
}
## 25-dimensional box constrained
#optim(rep(3, 25), flb,lower = rep(2, 25), upper = rep(4, 25),method = "L-BFGS-B") 
res_box <- optim_ARS(rep(3, 25), flb,lower = rep(2, 25), upper = rep(4, 25))

#> It.   5 | OFV = 3460
#> It.  10 | OFV = 2360.54
#> It.  15 | OFV = 2360.54
#> It.  20 | OFV = 2360.54
#> It.  25 | OFV = 2360.54
#> It.  30 | OFV = 2317.69
#> It.  35 | OFV = 2317.69
#> It.  40 | OFV = 2317.69
#> It.  45 | OFV = 2126.89
#> It.  50 | OFV = 2090.96
#> It.  55 | OFV = 1878.54
#> It.  60 | OFV = 1796.39
#> It.  65 | OFV = 1794.85
#> It.  70 | OFV = 1794.85
#> It.  75 | OFV = 1794.85
#> It.  80 | OFV = 1794.85
#> It.  85 | OFV = 1794.85
#> It.  90 | OFV = 1794.85
#> It.  95 | OFV = 1794.85
#> It. 100 | OFV = 1794.85
#> It. 105 | OFV = 1794.85
#> It. 110 | OFV = 1635.22
#> It. 115 | OFV = 1635.22
#> It. 120 | OFV = 1635.22
#> It. 125 | OFV = 1635.22
#> It. 130 | OFV = 1635.22
#> It. 135 | OFV = 1635.22
#> It. 140 | OFV = 1588.1
#> It. 145 | OFV = 1483.85
#> It. 150 | OFV = 1483.85
#> It. 155 | OFV = 1483.85
#> It. 160 | OFV = 1406.76
#> It. 165 | OFV = 1406.76
#> It. 170 | OFV = 1369.98
#> It. 175 | OFV = 1369.98
#> It. 180 | OFV = 1369.98
#> It. 185 | OFV = 1223.37
#> It. 190 | OFV = 1223.37
#> It. 195 | OFV = 1223.37
#> It. 200 | OFV = 1223.37
#> It. 205 | OFV = 1223.37
#> It. 210 | OFV = 1223.37
#> It. 215 | OFV = 1223.37
#> It. 220 | OFV = 1223.37
#> It. 225 | OFV = 1223.37
#> It. 230 | OFV = 1223.37
#> It. 235 | OFV = 1223.37
#> It. 240 | OFV = 1223.37
#> It. 245 | OFV = 1071.66
#> It. 250 | OFV = 991.893
#> It. 255 | OFV = 991.893
#> It. 260 | OFV = 897.197
#> It. 265 | OFV = 897.197
#> It. 270 | OFV = 847.195
#> It. 275 | OFV = 821.617
#> It. 280 | OFV = 730.11
#> It. 285 | OFV = 730.11
#> It. 290 | OFV = 730.11
#> It. 295 | OFV = 724.875
#> It. 300 | OFV = 724.875
#> It. 305 | OFV = 699.144
#> It. 310 | OFV = 699.144
#> It. 315 | OFV = 683.286
#> It. 320 | OFV = 667.358
#> It. 325 | OFV = 667.358
#> It. 330 | OFV = 667.358
#> It. 335 | OFV = 667.358
#> It. 340 | OFV = 667.358
#> It. 345 | OFV = 667.358
#> It. 350 | OFV = 667.358
#> It. 355 | OFV = 667.358
#> It. 360 | OFV = 667.358
#> It. 365 | OFV = 667.358
#> It. 370 | OFV = 667.358
#> It. 375 | OFV = 632.358
#> It. 380 | OFV = 632.358
#> It. 385 | OFV = 628.904
#> It. 390 | OFV = 628.904
#> It. 395 | OFV = 628.904
#> It. 400 | OFV = 628.904
#> 
#> Total iterations: 400 
#> Elapsed time: 0.248 seconds.
#> 
#> Final OFV =  628.9043 
#> Parameters: 2.34653 2 2.17345 2.054583 2 2.254394 2 2.014612 2 2.082373 2.176853 2.096385 2.033082 2.150576 2 2.472583 2.289041 2.126803 2.30252 2.055851 2.253626 2.545513 2 2.015903 3.613929 
#> 

## Combinatorial optimization: Traveling salesman problem
eurodistmat <- as.matrix(eurodist)

distance <- function(sq) {  # Target function
  sq2 <- embed(sq, 2)
  sum(eurodistmat[cbind(sq2[,2], sq2[,1])])
}

genseq <- function(sq) {  # Generate new candidate sequence
  idx <- seq(2, NROW(eurodistmat)-1)
  changepoints <- sample(idx, size = 2, replace = FALSE)
  tmp <- sq[changepoints[1]]
  sq[changepoints[1]] <- sq[changepoints[2]]
  sq[changepoints[2]] <- tmp
  sq
}

sq <- c(1:nrow(eurodistmat), 1)  # Initial sequence: alphabetic
res3 <- optim_ARS(sq,distance,generator=genseq) # Near optimum distance around 12842

#> It.   5 | OFV = 29012
#> It.  10 | OFV = 27791
#> It.  15 | OFV = 27791
#> It.  20 | OFV = 21841
#> It.  25 | OFV = 21795
#> It.  30 | OFV = 21795
#> It.  35 | OFV = 21795
#> It.  40 | OFV = 21795
#> It.  45 | OFV = 21292
#> It.  50 | OFV = 20619
#> It.  55 | OFV = 20619
#> It.  60 | OFV = 20619
#> It.  65 | OFV = 20619
#> It.  70 | OFV = 20619
#> It.  75 | OFV = 20619
#> It.  80 | OFV = 20361
#> It.  85 | OFV = 20279
#> It.  90 | OFV = 20279
#> It.  95 | OFV = 20279
#> It. 100 | OFV = 20279
#> It. 105 | OFV = 20279
#> It. 110 | OFV = 20279
#> It. 115 | OFV = 19698
#> It. 120 | OFV = 19587
#> It. 125 | OFV = 18416
#> It. 130 | OFV = 18193
#> It. 135 | OFV = 18193
#> It. 140 | OFV = 18193
#> It. 145 | OFV = 18193
#> It. 150 | OFV = 17945
#> It. 155 | OFV = 17945
#> It. 160 | OFV = 17945
#> It. 165 | OFV = 17945
#> It. 170 | OFV = 17723
#> It. 175 | OFV = 17723
#> It. 180 | OFV = 17723
#> It. 185 | OFV = 17723
#> It. 190 | OFV = 17344
#> It. 195 | OFV = 17344
#> It. 200 | OFV = 16340
#> It. 205 | OFV = 16340
#> It. 210 | OFV = 16340
#> It. 215 | OFV = 16340
#> It. 220 | OFV = 16340
#> It. 225 | OFV = 16340
#> It. 230 | OFV = 16340
#> It. 235 | OFV = 16340
#> It. 240 | OFV = 16340
#> It. 245 | OFV = 16109
#> It. 250 | OFV = 16097
#> It. 255 | OFV = 16097
#> It. 260 | OFV = 16097
#> It. 265 | OFV = 16097
#> It. 270 | OFV = 16097
#> It. 275 | OFV = 16097
#> It. 280 | OFV = 16097
#> It. 285 | OFV = 15884
#> It. 290 | OFV = 15884
#> It. 295 | OFV = 15884
#> It. 300 | OFV = 15884
#> It. 305 | OFV = 15884
#> It. 310 | OFV = 15884
#> It. 315 | OFV = 15884
#> It. 320 | OFV = 15884
#> It. 325 | OFV = 15884
#> It. 330 | OFV = 15884
#> It. 335 | OFV = 15884
#> It. 340 | OFV = 15884
#> It. 345 | OFV = 15884
#> It. 350 | OFV = 15884
#> It. 355 | OFV = 15884
#> It. 360 | OFV = 15884
#> It. 365 | OFV = 15884
#> It. 370 | OFV = 15884
#> It. 375 | OFV = 15884
#> It. 380 | OFV = 15884
#> It. 385 | OFV = 15884
#> It. 390 | OFV = 15884
#> It. 395 | OFV = 15884
#> It. 400 | OFV = 15884
#> 
#> Total iterations: 400 
#> Elapsed time: 0.281 seconds.
#> 
#> Final OFV =  15884 
#> Parameters: 1 6 18 5 4 10 20 7 11 3 13 12 9 14 2 15 8 16 19 17 21 1 
#> 

## Not run: ------------------------------------
#  
#   # plot of initial sequence
#   # rotate for conventional orientation
#   loc <- -cmdscale(eurodist, add = TRUE)$points
#   x <- loc[,1]; y <- loc[,2]
#   s <- seq_len(nrow(eurodistmat))
#   tspinit <- loc[sq,]
#   
#   plot(x, y, type = "n", asp = 1, xlab = "", ylab = "",
#        main = paste("Initial sequence of traveling salesman problem\n",
#                     "Distance =",distance(sq)), axes = FALSE)
#   arrows(tspinit[s,1], tspinit[s,2], tspinit[s+1,1], tspinit[s+1,2],
#          angle = 10, col = "green")
#   text(x, y, labels(eurodist), cex = 0.8)
#   
#   # plot of final sequence from optim_ARS
#   tspres <- loc[res3$par,]
#   plot(x, y, type = "n", asp = 1, xlab = "", ylab = "",
#        main = paste("optim_ARS() 'solving' traveling salesman problem\n",
#                     "Distance =",distance(c(1,res3$par,1))),axes = FALSE)
#   arrows(tspres[s,1], tspres[s,2], tspres[s+1,1], tspres[s+1,2],
#          angle = 10, col = "red")
#   text(x, y, labels(eurodist), cex = 0.8)
#   
#   # using optim
#   set.seed(123) # chosen to get a good soln relatively quickly
#   (res4 <- optim(sq, distance, genseq, method = "SANN",
#                  control = list(maxit = 30000, temp = 2000, trace = TRUE,
#                                 REPORT = 500))) 
#   
#   tspres <- loc[res4$par,]
#   plot(x, y, type = "n", asp = 1, xlab = "", ylab = "",
#        main = paste("optim() 'solving' traveling salesman problem\n",
#                     "Distance =",distance(res4$par)),axes = FALSE)
#   arrows(tspres[s,1], tspres[s,2], tspres[s+1,1], tspres[s+1,2],
#          angle = 10, col = "red")
#   text(x, y, labels(eurodist), cex = 0.8)
## ---------------------------------------------  

# one-dimensional function
## Not run: ------------------------------------
#  
#   f <- function(x)  abs(x)+cos(x)
#   res5 <- optim_ARS(-20,f,lower=-20, upper=20)
#   
#   curve(f, -20, 20)
#   abline(v = res5$par, lty = 4,col="green")
## ---------------------------------------------  

# one-dimensional function
f <- function(x)  (x^2+x)*cos(x) # -10 < x < 10
res_max <- optim_ARS(0,f,lower=-10, upper=10,maximize=TRUE) # sometimes to local maxima

#> It.   5 | OFV = 25.3766
#> It.  10 | OFV = 34.8644
#> It.  15 | OFV = 34.8644
#> It.  20 | OFV = 34.8644
#> It.  25 | OFV = 34.8644
#> It.  30 | OFV = 34.8644
#> It.  35 | OFV = 34.8644
#> It.  40 | OFV = 35.1207
#> It.  45 | OFV = 35.1207
#> It.  50 | OFV = 35.1207
#> It.  55 | OFV = 35.1207
#> It.  60 | OFV = 35.1207
#> It.  65 | OFV = 35.1207
#> It.  70 | OFV = 35.1207
#> It.  75 | OFV = 35.1207
#> It.  80 | OFV = 35.1207
#> It.  85 | OFV = 35.1207
#> It.  90 | OFV = 35.1207
#> It.  95 | OFV = 35.1207
#> It. 100 | OFV = 35.1207
#> It. 105 | OFV = 35.1207
#> It. 110 | OFV = 35.1207
#> It. 115 | OFV = 35.1207
#> It. 120 | OFV = 35.1207
#> It. 125 | OFV = 35.1207
#> It. 130 | OFV = 35.1207
#> It. 135 | OFV = 35.1207
#> It. 140 | OFV = 35.1207
#> It. 145 | OFV = 35.1207
#> It. 150 | OFV = 35.1207
#> It. 155 | OFV = 35.1207
#> It. 160 | OFV = 35.1207
#> It. 165 | OFV = 35.1207
#> It. 170 | OFV = 35.1207
#> It. 175 | OFV = 35.1207
#> It. 180 | OFV = 35.1207
#> It. 185 | OFV = 35.1207
#> It. 190 | OFV = 35.1207
#> It. 195 | OFV = 35.1207
#> It. 200 | OFV = 35.1207
#> It. 205 | OFV = 35.1207
#> It. 210 | OFV = 35.1207
#> It. 215 | OFV = 35.1207
#> It. 220 | OFV = 35.1207
#> It. 225 | OFV = 35.1207
#> It. 230 | OFV = 35.1207
#> It. 235 | OFV = 35.1207
#> Maximum number of identical optimal values reached (max_run=200), optimization stopped.
#> 
#> Total iterations: 236 
#> Elapsed time: 0.135 seconds.
#> 
#> Final OFV =  35.12066 
#> Parameters: -6.602795 
#> 

## Not run: ------------------------------------
#  
#   res_min <- optim_ARS(0,f,lower=-10, upper=10) # sometimes to local minima
#   
#   curve(f, -10, 10)
#   abline(v = res_min$par, lty = 4,col="green")
#   abline(v = res_max$par, lty = 4,col="red")
## ---------------------------------------------

# two-dimensional Rastrigin function
#It has a global minimum at f(x) = f(0) = 0.
## Not run: ------------------------------------
#  
#   Rastrigin <- function(x1, x2){
#     20 + x1^2 + x2^2 - 10*(cos(2*pi*x1) + cos(2*pi*x2))
#   }
#   
#   
#   x1 <- x2 <- seq(-5.12, 5.12, by = 0.1)
#   z <- outer(x1, x2, Rastrigin)
#   
#   res6 <- optim_ARS(c(-4,4),function(x) Rastrigin(x[1], x[2]),lower=-5.12, upper=5.12)
#   
#   # color scale
#   nrz <- nrow(z)
#   ncz <- ncol(z)
#   jet.colors <-
#     colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan",
#                        "#7FFF7F", "yellow", "#FF7F00", "red", "#7F0000"))
#   # Generate the desired number of colors from this palette
#   nbcol <- 100
#   color <- jet.colors(nbcol)
#   # Compute the z-value at the facet centres
#   zfacet <- z[-1, -1] + z[-1, -ncz] + z[-nrz, -1] + z[-nrz, -ncz]
#   # Recode facet z-values into color indices
#   facetcol <- cut(zfacet, nbcol)
#   persp(x1, x2, z, col = color[facetcol], phi = 30, theta = 30)
#   filled.contour(x1, x2, z, color.palette = jet.colors)
## ---------------------------------------------

## Parallel computation  
## works better when each evaluation takes longer
## here we have added extra time to the computations
## just to show that it works
## Not run: ------------------------------------
#  
#   res7 <- optim_ARS(c(-4,4),function(x){Sys.sleep(0.01); Rastrigin(x[1], x[2])},
#                     lower=-5.12, upper=5.12)
#   res8 <- optim_ARS(c(-4,4),function(x){Sys.sleep(0.01); Rastrigin(x[1], x[2])},
#                     lower=-5.12, upper=5.12,parallel = T)
#   res9 <- optim_ARS(c(-4,4),function(x){Sys.sleep(0.01); Rastrigin(x[1], x[2])},
#                     lower=-5.12, upper=5.12,parallel = T,parallel_type = "snow")
## ---------------------------------------------