Function written to match MATLAB's randn function

Generate random samples from a standardized normal distribution and return in matrix form.

randn(dim1, dim2 = NULL)

Arguments

dim1
The dimension of the matrix (if square), otherwise the number of rows.

dim2
The number of colums, if different from the number of rows.

Value

Matrix of random generated samples.

See also

Examples

randn(2,3)

#>            [,1]      [,2]       [,3]
#> [1,]  2.0646567 0.6338806  0.2501545
#> [2,] -0.1495222 0.4223444 -1.0401353

randn(5)

#>            [,1]         [,2]       [,3]         [,4]       [,5]
#> [1,] -0.0287149 -0.008765541 -0.6013227 -2.384946179 -1.6689696
#> [2,]  1.6451727  0.745206181 -1.4749807 -0.132497064  0.3094594
#> [3,]  0.2739507 -0.010497841 -0.6949729  1.052327238 -1.1409615
#> [4,]  0.7140061  0.215483199  0.4710895  0.021693757 -0.1954559
#> [5,] -1.4396878  1.105622658  2.3007945  0.002315637  0.9673516